@@ -1,11 +1,10 @@
 {
-  "name": "v-tour",
+  "name": "i-hack-v4",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
@@ -13,58 +12,49 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^5.2.11",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
+    "@angular/animations": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
     "@angular/fire": "^5.0.2",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "angular-background-image": "^0.1.2",
-    "angular-bootstrap-md": "^6.2.3",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@auth0/angular-jwt": "^2.0.0",
     "angular2-flash-messages": "^3.0.1",
     "angular2-jwt": "^0.2.3",
-    "angularx-qrcode": "^1.0.3",
-    "body-parser": "^1.18.3",
-    "bootstrap": "^4.1.1",
-    "core-js": "^2.4.1",
-    "cors": "^2.8.4",
-    "dateformat": "^3.0.3",
-    "express": "^4.16.3",
+    "angularx-qrcode": "^1.2.4",
+    "bootstrap": "^4.1.3",
+    "core-js": "^2.5.4",
     "firebase": "^5.5.3",
-    "google-maps-ng2": "^1.0.10",
     "jquery": "^3.3.1",
-    "mongodb": "^2.2.33",
-    "mongoose": "^5.2.9",
     "ngx-bootstrap": "^3.0.1",
-    "ngx-google-map": "^1.0.0",
-    "popper.js": "^1.14.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "popper.js": "^1.14.4",
+    "rxjs": "^6.0.0",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "~6.1.5",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.0.0",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.0",
-    "ts-node": "~4.1.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2"
   }
 }
