@@ -1,7 +1,11 @@
 import { Injectable } from '@angular/core';
 import { Http,Headers } from '@angular/http';
-import 'rxjs/add/operator/map';
-import { tokenNotExpired } from 'angular2-jwt';
+import { JwtHelperService } from '@auth0/angular-jwt';
+
+// 
+//import 'rxjs/add/operator/filter';
+import { map } from 'rxjs/operators';
+//import { tokenNotExpired } from 'angular2-jwt';
 
 @Injectable()
 export class AuthService {
@@ -10,7 +14,8 @@ export class AuthService {
   authtoken:any;
 
   constructor(
-    private http:Http
+    private http:Http,
+    public jwtHelper: JwtHelperService
   ) { }
 
   registerUser(user){
@@ -19,14 +24,14 @@ export class AuthService {
     let headers = new Headers();
    headers.append('Content-Type','application/json');
 
-    return this.http.post("http://localhost:3000/user/register",user,{headers:headers}).map(res=>res.json());
+    return this.http.post("http://localhost:3000/user/register",user,{headers:headers}).pipe(map(res=>res.json()));
   }
 
   loginUser(user){
     let headers = new Headers();
    headers.append('Content-Type','application/json');
 
-    return this.http.post("http://localhost:4201/login",user,{headers:headers}).map(res=>res.json());
+    return this.http.post("http://localhost:4201/login",user,{headers:headers}).pipe(map(res=>res.json()));
   }
 
   storeData(token,userData){
@@ -41,7 +46,7 @@ export class AuthService {
     let headers = new Headers();
     headers.append('Authorization',this.authtoken);
     headers.append('Content-Type','application/json');
-     return this.http.get("http://localhost:4201/profile",{headers:headers}).map(res=>res.json());  
+     return this.http.get("http://localhost:4201/profile",{headers:headers}).pipe(map(res=>res.json()));  
   }
 
   fetchToken(){
@@ -55,8 +60,9 @@ export class AuthService {
     localStorage.clear();
   }
 
-  loggedIn() {
-    return tokenNotExpired();
+  loggedIn():boolean {
+    const token = localStorage.getItem('token');
+    return !this.jwtHelper.isTokenExpired(token);
   }
   
 }
