@@ -1,17 +1,10 @@
 import { Component } from '@angular/core';
 
-
-
-
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-  
-  title: string = 'My first AGM project';
-  lat: number = 51.678418;
-  lng: number = 7.809007;
-  
+  title = 'iHackV4';
 }
