@@ -8,21 +8,19 @@ import { Observable } from 'rxjs';
   styleUrls: ['./map.component.css']
 })
 export class MapComponent implements OnInit {
-
-  title: string = 'My first AGM project';
-  lat: number = 51.678418;
-  lng: number = 7.809007;
-  qlng: any;
-  qlat: any;
+  
+  lat: number = 6.9021297;
+  lng: number = 79.8610308;
 
   items: Observable<any>;
+
   
   constructor(private db: AngularFireDatabase) { 
-    this.items = db.object('position').valueChanges();
-    this.qlng = db.object('position/longitude').valueChanges();
-    this.qlat = db.object('position/longitude').valueChanges();
-    
-    console.log("sdfsadf"+JSON.stringify(this.items));
+    db.object('/position').valueChanges()
+    .subscribe(res => {
+      this.lat = res.latitude;
+      this.lng = res.longitude;
+    });
   }
 
   ngOnInit() {
